Refactoring of structure and started to implement the tasks from nick
