+++ packages/web3-eth-ens/src/config.js
@@ -1,0 +1,6 @@
+"use strict";
+
+module.exports = {
+	main: "0x314159265dD8dbb310642f98f50C066173C1259b",
+	ropsten: "0x112234455c3a32fd11230c42e7bccd4a84e02010"
+};
