+++ packages/web3-eth-ens/src/contracts/ENSResolver.js
@@ -1,0 +1,5 @@
+"use strict";
+
+var ENSResolver = function() {
+
+}
