+++ packages/web3-eth-ens/src/index.js
@@ -22,662 +22,4 @@
  */
 
 "use strict";
-
-var ENS_ADDRESSES = {
-    main: "0x314159265dD8dbb310642f98f50C066173C1259b",
-    ropsten: "0x112234455c3a32fd11230c42e7bccd4a84e02010"
-};
-
-var ENS_ABI = [
-  {
-    "constant": true,
-    "inputs": [
-      {
-        "name": "node",
-        "type": "bytes32"
-      }
-    ],
-    "name": "resolver",
-    "outputs": [
-      {
-        "name": "",
-        "type": "address"
-      }
-    ],
-    "payable": false,
-    "type": "function"
-  },
-  {
-    "constant": true,
-    "inputs": [
-      {
-        "name": "node",
-        "type": "bytes32"
-      }
-    ],
-    "name": "owner",
-    "outputs": [
-      {
-        "name": "",
-        "type": "address"
-      }
-    ],
-    "payable": false,
-    "type": "function"
-  },
-  {
-    "constant": false,
-    "inputs": [
-      {
-        "name": "node",
-        "type": "bytes32"
-      },
-      {
-        "name": "label",
-        "type": "bytes32"
-      },
-      {
-        "name": "owner",
-        "type": "address"
-      }
-    ],
-    "name": "setSubnodeOwner",
-    "outputs": [],
-    "payable": false,
-    "type": "function"
-  },
-  {
-    "constant": false,
-    "inputs": [
-      {
-        "name": "node",
-        "type": "bytes32"
-      },
-      {
-        "name": "ttl",
-        "type": "uint64"
-      }
-    ],
-    "name": "setTTL",
-    "outputs": [],
-    "payable": false,
-    "type": "function"
-  },
-  {
-    "constant": true,
-    "inputs": [
-      {
-        "name": "node",
-        "type": "bytes32"
-      }
-    ],
-    "name": "ttl",
-    "outputs": [
-      {
-        "name": "",
-        "type": "uint64"
-      }
-    ],
-    "payable": false,
-    "type": "function"
-  },
-  {
-    "constant": false,
-    "inputs": [
-      {
-        "name": "node",
-        "type": "bytes32"
-      },
-      {
-        "name": "resolver",
-        "type": "address"
-      }
-    ],
-    "name": "setResolver",
-    "outputs": [],
-    "payable": false,
-    "type": "function"
-  },
-  {
-    "constant": false,
-    "inputs": [
-      {
-        "name": "node",
-        "type": "bytes32"
-      },
-      {
-        "name": "owner",
-        "type": "address"
-      }
-    ],
-    "name": "setOwner",
-    "outputs": [],
-    "payable": false,
-    "type": "function"
-  },
-  {
-    "anonymous": false,
-    "inputs": [
-      {
-        "indexed": true,
-        "name": "node",
-        "type": "bytes32"
-      },
-      {
-        "indexed": false,
-        "name": "owner",
-        "type": "address"
-      }
-    ],
-    "name": "Transfer",
-    "type": "event"
-  },
-  {
-    "anonymous": false,
-    "inputs": [
-      {
-        "indexed": true,
-        "name": "node",
-        "type": "bytes32"
-      },
-      {
-        "indexed": true,
-        "name": "label",
-        "type": "bytes32"
-      },
-      {
-        "indexed": false,
-        "name": "owner",
-        "type": "address"
-      }
-    ],
-    "name": "NewOwner",
-    "type": "event"
-  },
-  {
-    "anonymous": false,
-    "inputs": [
-      {
-        "indexed": true,
-        "name": "node",
-        "type": "bytes32"
-      },
-      {
-        "indexed": false,
-        "name": "resolver",
-        "type": "address"
-      }
-    ],
-    "name": "NewResolver",
-    "type": "event"
-  },
-  {
-    "anonymous": false,
-    "inputs": [
-      {
-        "indexed": true,
-        "name": "node",
-        "type": "bytes32"
-      },
-      {
-        "indexed": false,
-        "name": "ttl",
-        "type": "uint64"
-      }
-    ],
-    "name": "NewTTL",
-    "type": "event"
-  }
-];
-
-var RESOLVER_ABI = [
-  {
-    "constant": true,
-    "inputs": [
-      {
-        "name": "interfaceID",
-        "type": "bytes4"
-      }
-    ],
-    "name": "supportsInterface",
-    "outputs": [
-      {
-        "name": "",
-        "type": "bool"
-      }
-    ],
-    "payable": false,
-    "type": "function"
-  },
-  {
-    "constant": true,
-    "inputs": [
-      {
-        "name": "node",
-        "type": "bytes32"
-      },
-      {
-        "name": "contentTypes",
-        "type": "uint256"
-      }
-    ],
-    "name": "ABI",
-    "outputs": [
-      {
-        "name": "contentType",
-        "type": "uint256"
-      },
-      {
-        "name": "data",
-        "type": "bytes"
-      }
-    ],
-    "payable": false,
-    "type": "function"
-  },
-  {
-    "constant": false,
-    "inputs": [
-      {
-        "name": "node",
-        "type": "bytes32"
-      },
-      {
-        "name": "x",
-        "type": "bytes32"
-      },
-      {
-        "name": "y",
-        "type": "bytes32"
-      }
-    ],
-    "name": "setPubkey",
-    "outputs": [],
-    "payable": false,
-    "type": "function"
-  },
-  {
-    "constant": true,
-    "inputs": [
-      {
-        "name": "node",
-        "type": "bytes32"
-      }
-    ],
-    "name": "content",
-    "outputs": [
-      {
-        "name": "ret",
-        "type": "bytes32"
-      }
-    ],
-    "payable": false,
-    "type": "function"
-  },
-  {
-    "constant": true,
-    "inputs": [
-      {
-        "name": "node",
-        "type": "bytes32"
-      }
-    ],
-    "name": "addr",
-    "outputs": [
-      {
-        "name": "ret",
-        "type": "address"
-      }
-    ],
-    "payable": false,
-    "type": "function"
-  },
-  {
-    "constant": false,
-    "inputs": [
-      {
-        "name": "node",
-        "type": "bytes32"
-      },
-      {
-        "name": "contentType",
-        "type": "uint256"
-      },
-      {
-        "name": "data",
-        "type": "bytes"
-      }
-    ],
-    "name": "setABI",
-    "outputs": [],
-    "payable": false,
-    "type": "function"
-  },
-  {
-    "constant": true,
-    "inputs": [
-      {
-        "name": "node",
-        "type": "bytes32"
-      }
-    ],
-    "name": "name",
-    "outputs": [
-      {
-        "name": "ret",
-        "type": "string"
-      }
-    ],
-    "payable": false,
-    "type": "function"
-  },
-  {
-    "constant": false,
-    "inputs": [
-      {
-        "name": "node",
-        "type": "bytes32"
-      },
-      {
-        "name": "name",
-        "type": "string"
-      }
-    ],
-    "name": "setName",
-    "outputs": [],
-    "payable": false,
-    "type": "function"
-  },
-  {
-    "constant": false,
-    "inputs": [
-      {
-        "name": "node",
-        "type": "bytes32"
-      },
-      {
-        "name": "hash",
-        "type": "bytes32"
-      }
-    ],
-    "name": "setContent",
-    "outputs": [],
-    "payable": false,
-    "type": "function"
-  },
-  {
-    "constant": true,
-    "inputs": [
-      {
-        "name": "node",
-        "type": "bytes32"
-      }
-    ],
-    "name": "pubkey",
-    "outputs": [
-      {
-        "name": "x",
-        "type": "bytes32"
-      },
-      {
-        "name": "y",
-        "type": "bytes32"
-      }
-    ],
-    "payable": false,
-    "type": "function"
-  },
-  {
-    "constant": false,
-    "inputs": [
-      {
-        "name": "node",
-        "type": "bytes32"
-      },
-      {
-        "name": "addr",
-        "type": "address"
-      }
-    ],
-    "name": "setAddr",
-    "outputs": [],
-    "payable": false,
-    "type": "function"
-  },
-  {
-    "inputs": [
-      {
-        "name": "ensAddr",
-        "type": "address"
-      }
-    ],
-    "payable": false,
-    "type": "constructor"
-  },
-  {
-    "anonymous": false,
-    "inputs": [
-      {
-        "indexed": true,
-        "name": "node",
-        "type": "bytes32"
-      },
-      {
-        "indexed": false,
-        "name": "a",
-        "type": "address"
-      }
-    ],
-    "name": "AddrChanged",
-    "type": "event"
-  },
-  {
-    "anonymous": false,
-    "inputs": [
-      {
-        "indexed": true,
-        "name": "node",
-        "type": "bytes32"
-      },
-      {
-        "indexed": false,
-        "name": "hash",
-        "type": "bytes32"
-      }
-    ],
-    "name": "ContentChanged",
-    "type": "event"
-  },
-  {
-    "anonymous": false,
-    "inputs": [
-      {
-        "indexed": true,
-        "name": "node",
-        "type": "bytes32"
-      },
-      {
-        "indexed": false,
-        "name": "name",
-        "type": "string"
-      }
-    ],
-    "name": "NameChanged",
-    "type": "event"
-  },
-  {
-    "anonymous": false,
-    "inputs": [
-      {
-        "indexed": true,
-        "name": "node",
-        "type": "bytes32"
-      },
-      {
-        "indexed": true,
-        "name": "contentType",
-        "type": "uint256"
-      }
-    ],
-    "name": "ABIChanged",
-    "type": "event"
-  },
-  {
-    "anonymous": false,
-    "inputs": [
-      {
-        "indexed": true,
-        "name": "node",
-        "type": "bytes32"
-      },
-      {
-        "indexed": false,
-        "name": "x",
-        "type": "bytes32"
-      },
-      {
-        "indexed": false,
-        "name": "y",
-        "type": "bytes32"
-      }
-    ],
-    "name": "PubkeyChanged",
-    "type": "event"
-  }
-];
-
-var namehash = require('eth-ens-namehash');
-var _ = require('underscore');
-var Contract = require('web3-eth-contract');
-
-/**
- * A wrapper around the ENS registry contract.
- *
- * @method ENSRegistry
- * @constructor
- * @param {Object} eth
- */
-var ENSRegistry = function ENSRegistry(ens) {
-    this._ens = ens;
-    this._contract = ens.checkNetwork().then(function (address) {
-        return new Contract(ENS_ABI, address);
-    });
-};
-
-/**
- * Returns the address of the owner of an ENS name.
- *
- * @method owner
- * @constructor
- * @param {string} name
- * @param {Function} optional callback
- * @return {Function} a promise
- */
-ENSRegistry.prototype.owner = function owner(name, callback) {
-    return this._contract
-    .then(function(contract) {
-        return contract.methods.owner(namehash.hash(name)).call();
-    }, function(err) {
-        if(_.isFunction(callback)) callback(err);
-        throw err;
-    });
-};
-
-/**
- * Returns the resolver contract associated with a name.
- *
- * @method owner
- * @constructor
- * @param {string} name
- * @param {Function} optional callback
- * @return {Function} a promise that yields a resolver contract instance
- */
-ENSRegistry.prototype.resolver = function resolver(name, callback) {
-    var node = namehash.hash(name);
-
-    return this._contract
-    .then(function(contract) {
-        return contract.methods.resolver(node).call();
-    })
-    .then(function(addr) {
-        var contract = new Contract(RESOLVER_ABI, addr);
-        _.each(contract.methods, function(method, name) {
-            if(name === 'supportsInterface') return;
-            contract.methods[name] = _.partial(method, node);
-        });
-
-        if(_.isFunction(callback)) callback(null, contract);
-        return contract;
-    }, function(err) {
-        if(_.isFunction(callback)) callback(err);
-        throw err;
-    });
-};
-
-/**
- * Constructs a new ENS instance.
- *
- * @method ENS
- * @constructor
- * @param {Object} eth
- */
-var ENS = function ENS(eth) {
-    this._eth = eth;
-};
-
-/**
- * Checks that the current network supports ENS and is synced. Throws an error if not.
- *
- * @method checkNetwork
- * @param {Function} callback
- * @return The address of the ENS registry, if present on this network and if
- *         the node is synced.
- */
-ENS.prototype.checkNetwork = function checkNetwork(callback) {
-    var eth = this._eth;
-    return eth.getBlock('latest').then(function(block) {
-        var headAge = new Date() / 1000 - block.timestamp;
-        if(headAge > 3600) {
-            throw new Error("Network not synced; last block was " + headAge + " seconds ago");
-        }
-        return eth.net.getNetworkType();
-    })
-    .then(function(networkType) {
-        var addr = ENS_ADDRESSES[networkType];
-        if(addr === undefined) {
-            throw new Error("ENS is not supported on network " + networkType);
-        }
-        if(_.isFunction(callback)) callback(null, addr);
-        return addr;
-    })
-    .catch(function(err) {
-        if(_.isFunction(callback)) callback(err);
-        throw err;
-    });
-};
-
-Object.defineProperty(ENS.prototype, 'registry', {
-    get: function(){
-        return new ENSRegistry(this);
-    },
-    enumerable: true
-});
-
-/**
- * Gets the address record associated with a name.
- *
- * @method address
- * @param {string} name
- * @param {Function} optional callback
- * @return {Function} a promise
- */
-ENS.prototype.getAddress = function getAddress(name, callback) {
-    return this.registry.resolver(name).then(function(resolver) {
-        return resolver.methods.addr().call(callback);
-    }, function(err) {
-        if(_.isFunction(callback)) callback(err);
-        throw err;
-    });
-};
-
-module.exports = ENS;
+module.exports = require('./contracts/ENS');
